build module with rollup \o/
