+++ doc/global.html
@@ -1,0 +1,292 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>Global - Documentation</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <script src="scripts/prettify/prettify.js"></script>
+    <script src="scripts/prettify/lang-css.js"></script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="css/jsdoc-default.css">
+</head>
+<body>
+
+<input type="checkbox" id="nav-trigger" class="nav-trigger" />
+<label for="nav-trigger" class="navicon-button x">
+  <div class="navicon"></div>
+</label>
+
+<label for="nav-trigger" class="overlay"></label>
+
+<nav class="main-nav">
+    <header class="nav-header"><h2><a href="index.html">Home</a></h2></header><header class="nav-header"><h3><span>Classes</span></h3></header><ul><li><a href="module-holmes.html">holmes</a><ul class='members'><li class='nav-type-header'>Members</li><li data-type='member'><a href="module-holmes.html#elementsArray">elementsArray</a></li><li data-type='member'><a href="module-holmes.html#elementsLength">elementsLength</a></li><li data-type='member'><a href="module-holmes.html#hidden">hidden</a></li><li data-type='member'><a href="module-holmes.html#input">input</a></li><li data-type='member'><a href="module-holmes.html#options">options</a></li><li data-type='member'><a href="module-holmes.html#placeholderNode">placeholderNode</a></li><li data-type='member'><a href="module-holmes.html#searchString">searchString</a></li></ul><ul class='methods'><li class='nav-type-header'>Methods</li><li data-type='method'><a href="module-holmes.html#_mergeObj">_mergeObj</a></li><li data-type='method'><a href="module-holmes.html#clear">clear</a></li><li data-type='method'><a href="module-holmes.html#count">count</a></li><li data-type='method'><a href="module-holmes.html#start">start</a></li><li data-type='method'><a href="module-holmes.html#stop">stop</a></li><li data-type='method'><a href="module-holmes.html#~hideElement">hideElement</a></li><li data-type='method'><a href="module-holmes.html#~inputHandler">inputHandler</a></li><li data-type='method'><a href="module-holmes.html#~inputString">inputString</a></li><li data-type='method'><a href="module-holmes.html#~showElement">showElement</a></li></ul></li></ul><header class="nav-header"><h3><a href="global.html">Global</a></h3></header>
+</nav>
+
+<main id="main" class="main-content">
+    <section>
+    <h1 class="page-title">Global</h1>
+    
+
+    
+
+
+
+
+<section>
+
+<header>
+    
+        <h2 class="section-title">
+        
+        </h2>
+        
+    
+</header>
+
+<article>
+    <div class="container-overview">
+    
+        
+
+        
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+        
+    
+    </div>
+    
+
+    
+
+    
+
+     
+
+    
+    
+
+    
+
+    
+        <h3 class="subsection-title">Type Definitions</h3>
+
+        
+                
+<article class="card card__typedef" id="onChange">
+
+
+<section class="code-source">
+  <div class="link-box">
+    
+    <a href="main.js.html">main.js</a> <a href="main.js.html#line423">line 423</a>
+    
+  </div>
+</section>
+
+
+    
+      <h2 >onChange</h2>
+    
+    
+        
+
+        <h4 class="name"><span class="type-signature"></span>onChange<span class="signature">(element<span class="signature-attributes">opt</span>)</span><span class="type-signature"></span></h4>
+    
+
+    
+
+
+    
+    <div class="description">
+        Callback used for changes in item en list states.
+    </div>
+    
+
+    
+
+    
+
+    
+
+    
+        <h5>Parameters:</h5>
+        
+<div class="table-scroll">
+
+<table class="params">
+    <thead>
+    <tr>
+        
+        <th>Name</th>
+        
+
+        <th>Type</th>
+
+        
+        <th>Attributes</th>
+        
+
+        
+
+        <th class="last">Description</th>
+    </tr>
+    </thead>
+
+    <tbody>
+    
+
+        <tr>
+            
+                <td class="name"><code>element</code></td>
+            
+
+            <td class="type">
+            
+                
+<span class="__tag param-type param-type-object"><a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/object" target="_blank">object</a>
+
+</span>
+
+
+
+            
+            </td>
+
+            
+                <td class="attributes">
+                
+                    <span class="__tag attr attr-optional">optional</span><br>
+                
+
+                
+
+                
+                </td>
+            
+
+            
+
+            <td class="description last">Element affected by the event. This is the item found by
+  <code>onVisible</code> and <code>onHidden</code> and the placeholder
+  (or <code>undefined</code>) for <code>onEmpty</code> and
+  <code>onFound</code>.</td>
+        </tr>
+
+    
+    </tbody>
+</table>
+</div>
+
+    
+
+    
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    </article>
+
+            
+    
+
+    
+</article>
+
+</section>
+
+
+
+
+    </section>
+</main>
+
+
+<footer class="main-footer">
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.4.3</a> on Tue Nov 22 2016 18:57:26 GMT+0100 (CET)
+</footer>
+
+<script>prettyPrint();</script>
+<script src="scripts/linenumber.js"></script>
+</body>
+</html>
